+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 ### Version 10.2
+- Added JSON export for formula making
+
 ### Version 10.1
 - Added export to JSON for perfume types
 - Added export to JSON for categories