+++ releasenotes.md
@@ -1,3 +1,4 @@
 Whats New in v10.2
 --------------------------
+- Added JSON export for formula making
 - For full details please refer to the CHANGELOG